fix: use resized-output dims for thumbnails, CORS guard errors, dedup uploads
Address review-bot findings (gemini-code-assist + cubic):
- saveThumbnail (LocalFile + S3) now stores the actual resized WebP
  dimensions and byte size from sharp's toFile() OutputInfo, instead of
  sharp.metadata() which reports the *input* image.
- Media routes pass corsErrorResponse(CORS_HEADERS) so the OAuth guard's
  401/403/500 auth-failure responses carry the route's CORS headers.
- Extract the shared synchronous-upload handler (handleSyncMediaUpload) used
  by POST /api/v1/media and /api/v2/media to remove the duplicated body.
