+++ app/api/v2/media/route.ts
@@ -1,10 +1,11 @@
-import { OAuthGuardAnyScope } from '@/lib/services/guards/OAuthGuard'
-import { saveMedia } from '@/lib/services/medias'
-import { MediaSchema } from '@/lib/services/medias/types'
+import {
+  OAuthGuardAnyScope,
+  corsErrorResponse
+} from '@/lib/services/guards/OAuthGuard'
+import { handleSyncMediaUpload } from '@/lib/services/medias/handleSyncMediaUpload'
 import { Scope } from '@/lib/types/database/operations'
 import { HttpMethod } from '@/lib/utils/http-headers'
-import { logger } from '@/lib/utils/logger'
-import { ERROR_422, apiResponse, defaultOptions } from '@/lib/utils/response'
+import { defaultOptions } from '@/lib/utils/response'
 import { traceApiRoute } from '@/lib/utils/traceApiRoute'
 
 const CORS_HEADERS = [HttpMethod.enum.OPTIONS, HttpMethod.enum.POST]
@@ -21,44 +22,7 @@ export const POST = traceApiRoute(
   'uploadMediaV2',
   OAuthGuardAnyScope(
     [Scope.enum.write, Scope.enum['write:media']],
-    async (req, context) => {
-      try {
-        const { database, currentActor } = context
-        const form = await req.formData()
-        const media = MediaSchema.safeParse(Object.fromEntries(form.entries()))
-        if (!media.success) {
-          return apiResponse({
-            req,
-            allowedMethods: CORS_HEADERS,
-            data: ERROR_422,
-            responseStatusCode: 422
-          })
-        }
-
-        const response = await saveMedia(database, currentActor, media.data)
-        if (!response) {
-          return apiResponse({
-            req,
-            allowedMethods: CORS_HEADERS,
-            data: ERROR_422,
-            responseStatusCode: 422
-          })
-        }
-        return apiResponse({
-          req,
-          allowedMethods: CORS_HEADERS,
-          data: response
-        })
-      } catch (e) {
-        const nodeErr = e as NodeJS.ErrnoException
-        logger.error(nodeErr)
-        return apiResponse({
-          req,
-          allowedMethods: CORS_HEADERS,
-          data: ERROR_422,
-          responseStatusCode: 422
-        })
-      }
-    }
+    (req, context) => handleSyncMediaUpload(req, context, CORS_HEADERS),
+    { errorResponse: corsErrorResponse(CORS_HEADERS) }
   )
 )
